refactor: split ChatView.swift into MessageListView, ComposerSection, and ErrorBannerView
